// src/csv/formats/emars/types.ts
export type EmarsRow = {
  ssn: string;
  firstName: string;
  lastName: string;
  sex: "M" | "F";
  ethnicity: string;
  addressLine1: string;
  state: string;
  zip: string;
  workClass: string;
  appLevel: string;
  appPercent: string;
  straightTime1: string;
  straightTime2: string;
  straightTime3: string;
  straightTime4: string;
  straightTime5: string;
  straightTime6: string;
  straightTime7: string;
  overtime1: string;
  overtime2: string;
  overtime3: string;
  overtime4: string;
  overtime5: string;
  overtime6: string;
  overtime7: string;
  doubleTime1: string;
  doubleTime2: string;
  doubleTime3: string;
  doubleTime4: string;
  doubleTime5: string;
  doubleTime6: string;
  doubleTime7: string;
  doubleTimePayRate: string;
  overtimePayRate: string;
  straightTimePayRate: string;
  federalWitholding: string;
  fica: string;
  medicare: string;
  stateTax: string;
  localTax1: string;
  exemptions: string;
  grossPay: string;
  netPay: string;
};